pyproject.toml
[project]
name = "sn-gamestate"
version = "0.2.0"
description = "A Framework for football players identification"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "tracklab",
    "easyocr==1.7.1",
    "soccernet==0.1.55",
    "torch==2.5.0",
    "prtreid @ git+https://github.com/VlSomers/prtreid",
    "mmocr==1.0.1",
    "mmdet~=3.1.0",
    "openmim==0.3.9",
    "torchreid @ git+https://github.com/VlSomers/bpbreid",
    "lightning==2.0.9",
    "tracklab_calibration",
    "numpy",
    "supervision",
    "tensorrt-cu12",
    "tensorrt-lean-cu12",
    "tensorrt-dispatch-cu12",
    "tensorrt==10.7",
    "onnxruntime",
    "onnxruntime-gpu",
    "ultralytics",
    "inference",
    "loguru",
    "scikit-learn",
    "lap",
    "Cython",
    "h5py",
    "Pillow",
    "six",
    "scipy",
    "opencv-python",
    "matplotlib",
    "tb-nightly",
    "future",
    "yacs",
    "gdown",
    "flake8",
    "yapf",
    "isort==4.3.21",
    "imageio",
    "chardet",
    "tabulate",
    "pycocotools",
    "thop",
    "seaborn",
]

[project.entry-points.tracklab_plugin]
sn_gamestate = "sn_gamestate.config_finder:ConfigFinder"

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sn_gamestate*"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.rst", "*.yaml"]

[project.optional-dependencies]
dev = ["ipykernel==6.29.4"]

[tool.uv.sources]
tracklab-calibration = { path = "plugins/calibration", editable = true }
